Move all files outside desktop dir
The desktop dir was supposed to contain the files for the desktop
version of the app, in the hope that there would be other versions
and that they would be stored in this repo as well.

This does not seem like the best way to store things anymore. If
the app is ever extended to another platform, it will have a separate
repo.